Move IMGUI implementations of property drawers to separate file in obsolete
