+++ com.unity.cinemachine/Editor/Obsolete/ObsoletePropertyDrawers.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 34a7bac4e42d2e04b941e57253dcd670
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
