+++ com.unity.cinemachine/Editor/PropertyDrawers/GroupTargetPropertyDrawer.cs
@@ -1,4 +1,3 @@
-using UnityEngine;
 using UnityEditor;
 using UnityEngine.UIElements;
 using UnityEditor.UIElements;
@@ -6,35 +5,9 @@ using UnityEditor.UIElements;
 namespace Unity.Cinemachine.Editor
 {
     [CustomPropertyDrawer(typeof(CinemachineTargetGroup.Target))]
-    class GroupTargetPropertyDrawer : PropertyDrawer
+    partial class GroupTargetPropertyDrawer : PropertyDrawer
     {
         CinemachineTargetGroup.Target def = new();
-
-        public override void OnGUI(Rect rect, SerializedProperty property, GUIContent label)
-        {
-            float floatFieldWidth = EditorGUIUtility.singleLineHeight * 3.5f;
-
-            EditorGUI.BeginProperty(rect, GUIContent.none, property);
-
-            float oldWidth = EditorGUIUtility.labelWidth;
-            EditorGUIUtility.labelWidth = 1;
-
-            rect.width -= 2 * (floatFieldWidth + EditorGUIUtility.singleLineHeight);
-            var p = property.FindPropertyRelative(() => def.Object);
-            EditorGUI.PropertyField(rect, p, new GUIContent(" ", p.tooltip));
-
-            EditorGUIUtility.labelWidth = EditorGUIUtility.singleLineHeight;
-            rect.x += rect.width; rect.width = floatFieldWidth + EditorGUIUtility.singleLineHeight;
-            p = property.FindPropertyRelative(() => def.Weight);
-            EditorGUI.PropertyField(rect, p, new GUIContent(" ", p.tooltip));
-
-            rect.x += rect.width;
-            p = property.FindPropertyRelative(() => def.Radius);
-            EditorGUI.PropertyField(rect, p, new GUIContent(" ", p.tooltip));
-
-            EditorGUIUtility.labelWidth = oldWidth;
-            EditorGUI.EndProperty();
-        }
 
         public override VisualElement CreatePropertyGUI(SerializedProperty property)
         {
