+++ com.unity.cinemachine/Editor/PropertyDrawers/HideFoldoutPropertyDrawer.cs
@@ -1,23 +1,12 @@
 using UnityEditor;
 using UnityEditor.UIElements;
-using UnityEngine;
 using UnityEngine.UIElements;
 
 namespace Unity.Cinemachine.Editor
 {
     [CustomPropertyDrawer(typeof(HideFoldoutAttribute))]
-    class HideFoldoutPropertyDrawer : PropertyDrawer
+    partial class HideFoldoutPropertyDrawer : PropertyDrawer
     {
-        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
-        {
-            return InspectorUtility.PropertyHeightOfChidren(property);
-        }
-
-        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
-        {
-            InspectorUtility.DrawChildProperties(position, property);
-        }
-
         public override VisualElement CreatePropertyGUI(SerializedProperty property)
         {
             var ux = new VisualElement();
