+++ com.unity.cinemachine/Editor/PropertyDrawers/InputAxisPropertyDrawer.cs
@@ -1,4 +1,3 @@
-using UnityEngine;
 using UnityEditor;
 using UnityEngine.UIElements;
 using UnityEditor.UIElements;
@@ -6,86 +5,9 @@ using UnityEditor.UIElements;
 namespace Unity.Cinemachine.Editor
 {
     [CustomPropertyDrawer(typeof(InputAxis))]
-    class InputAxisWithNamePropertyDrawer : PropertyDrawer
+    partial class InputAxisWithNamePropertyDrawer : PropertyDrawer
     {
         InputAxis def = new (); // to access name strings
-
-        public override void OnGUI(Rect rect, SerializedProperty property, GUIContent label)
-        {
-            float height = EditorGUIUtility.singleLineHeight;
-            rect.height = height;
-
-            property.isExpanded = EditorGUI.Foldout(
-                new Rect(rect.x, rect.y, EditorGUIUtility.labelWidth - 2 * height, rect.height),
-                property.isExpanded, label, true);
-
-            if (property.isExpanded)
-            {
-                ++EditorGUI.indentLevel;
-
-                rect.y += height + EditorGUIUtility.standardVerticalSpacing;
-                EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => def.Value));
-
-                var flags = property.FindPropertyRelative(() => def.Restrictions).intValue;
-
-                var enabled = GUI.enabled;
-                GUI.enabled = (flags & (int)InputAxis.RestrictionFlags.RangeIsDriven) == 0;
-
-                rect.y += height + EditorGUIUtility.standardVerticalSpacing;
-                EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => def.Center));
-
-                rect.y += height + EditorGUIUtility.standardVerticalSpacing;
-                if ((flags & (int)InputAxis.RestrictionFlags.Momentary) != 0)
-                    EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => def.Range));
-                else
-                {
-                    InspectorUtility.MultiPropertyOnLine(
-                        rect, null,
-                        new [] {
-                                property.FindPropertyRelative(() => def.Range),
-                                property.FindPropertyRelative(() => def.Wrap)}, 
-                        new [] { GUIContent.none, null });
-                }
-
-                rect.y += height + EditorGUIUtility.standardVerticalSpacing;
-                if ((flags & (int)(InputAxis.RestrictionFlags.NoRecentering | InputAxis.RestrictionFlags.Momentary)) == 0)
-                    EditorGUI.PropertyField(rect, property.FindPropertyRelative(() => def.Recentering));
-
-                GUI.enabled = enabled;
-                --EditorGUI.indentLevel;
-            }
-            else
-            {
-                // Draw the input value on the same line as the foldout, for convenience
-                var valueProp = property.FindPropertyRelative(() => def.Value);
-
-                int oldIndent = EditorGUI.indentLevel;
-                float oldLabelWidth = EditorGUIUtility.labelWidth;
-
-                rect.x += EditorGUIUtility.labelWidth - 2 * EditorGUIUtility.singleLineHeight;
-                rect.width -= EditorGUIUtility.labelWidth - 2 * EditorGUIUtility.singleLineHeight;
-
-                EditorGUI.indentLevel = 0;
-                EditorGUIUtility.labelWidth = 2 * EditorGUIUtility.singleLineHeight;
-                EditorGUI.PropertyField(rect, valueProp, new GUIContent(" ", valueProp.tooltip));
-                EditorGUI.indentLevel = oldIndent;
-                EditorGUIUtility.labelWidth = oldLabelWidth;
-            }
-        }
-
-        public override float GetPropertyHeight(SerializedProperty property, GUIContent label)
-        {
-            var lineHeight = EditorGUIUtility.singleLineHeight + EditorGUIUtility.standardVerticalSpacing;
-            var height = lineHeight;
-            if (property != null && property.isExpanded)
-            {
-                height += 3 * lineHeight;
-                var flags = property.FindPropertyRelative(() => def.Restrictions).intValue;
-                if ((flags & (int)(InputAxis.RestrictionFlags.NoRecentering | InputAxis.RestrictionFlags.Momentary)) == 0)
-                    height += EditorGUI.GetPropertyHeight(property.FindPropertyRelative(() => def.Recentering));
-            }
-            return height - EditorGUIUtility.standardVerticalSpacing;
-        }
 
         public override VisualElement CreatePropertyGUI(SerializedProperty property)
         {
