+++ com.unity.cinemachine/Editor/PropertyDrawers/MinMaxRangeSliderPropertyDrawer.cs
@@ -1,4 +1,3 @@
-using System;
 using UnityEditor;
 using UnityEditor.UIElements;
 using UnityEngine;
@@ -7,37 +6,8 @@ using UnityEngine.UIElements;
 namespace Unity.Cinemachine.Editor
 {
     [CustomPropertyDrawer(typeof(MinMaxRangeSliderAttribute))]
-    class MinMaxRangeSliderPropertyDrawer : PropertyDrawer
+    partial class MinMaxRangeSliderPropertyDrawer : PropertyDrawer
     {
-        // IMGUI implementation must remain until no more IMGUI inspectors are using it
-        public override void OnGUI(Rect rect, SerializedProperty property, GUIContent label)
-        {
-            var a = attribute as MinMaxRangeSliderAttribute;
-            EditorGUI.BeginProperty(rect, label, property);
-            {
-                var v = property.vector2Value;
-
-                // The layout system breaks alignment when mixing inspector fields with custom layout'd
-                // fields as soon as a scrollbar is needed in the inspector, so we'll do the layout
-                // manually instead
-                const int kFloatFieldWidth = 50;
-                const int kSeparatorWidth = 5;
-                float indentOffset = EditorGUI.indentLevel * 15f;
-                var labelRect = new Rect(rect.x, rect.y, EditorGUIUtility.labelWidth - indentOffset, rect.height);
-                var floatFieldLeft = new Rect(labelRect.xMax, rect.y, kFloatFieldWidth + indentOffset, rect.height);
-                var sliderRect = new Rect(floatFieldLeft.xMax + kSeparatorWidth - indentOffset, rect.y, rect.width - labelRect.width - kFloatFieldWidth * 2 - kSeparatorWidth * 2, rect.height);
-                var floatFieldRight = new Rect(sliderRect.xMax + kSeparatorWidth - indentOffset, rect.y, kFloatFieldWidth + indentOffset, rect.height);
-
-                EditorGUI.PrefixLabel(labelRect, label);
-                v.x = EditorGUI.FloatField(floatFieldLeft, v.x);
-                EditorGUI.MinMaxSlider(sliderRect, ref v.x, ref v.y, a.Min, a.Max);
-                v.y = EditorGUI.FloatField(floatFieldRight, v.y);
-
-                property.vector2Value = v;
-            }
-            EditorGUI.EndProperty();
-        }
-
         public override VisualElement CreatePropertyGUI(SerializedProperty property)
         {
             var a = attribute as MinMaxRangeSliderAttribute;
